Log error when a thead or tbody tag is missing
When the debug option is enabled, this will report when floatThead can't find the thead or tbody. Currently it just doesn't work without any notification. 

The help page should probably also mention that a tbody is needed. Browsers and other similar plugins can handle tables without it no problem, so it can be quite a challenging problem to find.